refacto: changement de qq fonctions
